requirements.txt
# requirements.txt
numpy>=1.21.0
matplotlib>=3.4.0
opencv-python>=4.5.0
scipy>=1.7.0
Pillow>=8.3.0